% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_offnode.R
\name{add_offnode}
\alias{add_offnode}
\title{Off-variant-identification}
\usage{
add_offnode(blocklist, dataset, indi, nwindow, window_sequence, bp_map,
  off_node_minimum_blocklength = 10, off_node_minimum_blocksize = 5,
  raster = 5)
}
\arguments{
\item{blocklist}{blocklist created by blocklist_calculcation algorithm}

\item{dataset}{dataset which variant nr. for each window}

\item{indi}{number of haplotypes in the dataset}

\item{nwindow}{number of windows in the dataset}

\item{window_sequence}{sequence of predefined windows (default: NULL ;per row: start$snp, end$snp, length, length - merging_error, start$bp, end$bp)}

\item{bp_map}{vector of positions for each SNP in bp (default: NULL - all 0)}

\item{off_node_minimum_blocklength}{Minimum length of newly identified blocks (default: 10)}

\item{off_node_minimum_blocksize}{Minimum number of individuals in newly identified blocks (default: 5)}

\item{raster}{Raster-width in the identification step (default: 5; recommended to be lower than off_node_minimum_blocklength)}
}
\value{
haplotype library
}
\description{
Function to add off-variant-blocks
}
