% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_calculation.R
\name{block_calculation}
\alias{block_calculation}
\title{Mainfunction to calculate haplotype blocks}
\usage{
block_calculation(dhm, window_sequence = NULL, window_size = 20,
  merging_error = 1, node_min = 5, gap = 10, min_share = 0.975,
  off_lines = 2, min_similarity = 0.99, merge_closeblock = FALSE,
  max_diff_l = 1, max_diff_i = 1, min_majorblock = 5000,
  bp_map = NULL, window_anchor_gens = NULL, consider_nodes = TRUE,
  consider_edge = TRUE, edge_min = 5, subgroups = NULL,
  min_per_subgroup = 0, subgroup_exception = 0, consider_all = TRUE,
  save_allblock = TRUE, block_extending = TRUE,
  max_extending_diff = 1, extending_ratio = 20,
  min_majorblock_steps = 4, snp_extending = TRUE,
  max_extending_diff_snp = 0, extending_ratio_snp = Inf,
  major_snp_calculation = TRUE, off_node_addition = FALSE,
  off_node_minimum_blocklength = 10, off_node_minimum_blocksize = 5,
  raster = 5, at_least_one = TRUE, prefilter = FALSE, maf = 0,
  equal_remove = FALSE, big_output = FALSE, blockinfo_mode = 0,
  c_dhm_mode = TRUE, intersect_func = TRUE, fast_compiler = TRUE,
  max_groups = 0, recoding = FALSE, recoding_notneeded = FALSE,
  consider_multi = FALSE, multi_min = 5, blockinfo_mode_na = FALSE,
  actual_snp_weight = 5, na_snp_weight = 2, na_seq_weight = 0,
  weighting_length = 1, weighting_size = 1,
  recalculate_biggest = TRUE, target_coverage = NULL,
  max_iteration = 10, min_step_size = 25, target_stop = 0.001,
  multi_window_mode = FALSE, adaptive_mode = FALSE,
  developer_mode = FALSE, double_share = 1, early_remove = FALSE,
  node_min_early = NULL, min_reduction_cross = -Inf,
  min_reduction_neglet = -Inf, parallel_window = Inf,
  window_overlap = 0, window_cores = 1)
}
\arguments{
\item{dhm}{haploid SNP-dataset}

\item{window_sequence}{sequence of predefined windows (default: NULL ;per row: start$snp, end$snp, length, length - merging_error, start$bp, end$bp)}

\item{window_size}{size of each window in the algorithm (default: 20)}

\item{merging_error}{number of allowed errors per block (default: 1)}

\item{node_min}{minimum number of haplotypes per block (default: 5)}

\item{gap}{remove haplotypes in nodes adjacent to nodes with less than min_node haplotypes in it (default: 10 windows)}

\item{min_share}{minimum percentage of transition to the same block for extension (default: 0.95, step III)}

\item{off_lines}{minimum number of haplotypes to looose in the creation of a bigger block (default: 5)}

\item{min_similarity}{minimum rate of the same SNPs to be added to the block (default: 0.99)}

\item{merge_closeblock}{If TRUE merge adjecting block with similar haplotypes (default: FALSE)}

\item{max_diff_l}{maximum number of windows with different haplotypes inbetween (default: 1)}

\item{max_diff_i}{maximum number of individuals in only one of the two blocks (default: 1)}

\item{min_majorblock}{minimum of positions in the dataset a block is the biggest covering (default: 50)}

\item{bp_map}{vector of positions for each SNP in bp (default: NULL - all 0)}

\item{window_anchor_gens}{matrix to constructed window_sequence base on start/end points in bp (e.g. gen regions, per row: start, end)}

\item{consider_nodes}{Use nodes to identify blocks (default: TRUE)}

\item{consider_edge}{Use edges between nodes to identify blocks (default: TRUE)}

\item{edge_min}{minimum number of haplotypes per transition to use in consider_edge (default: 5)}

\item{subgroups}{possible subgroups to consider in the block identification process (default: NULL - list(1:indi))}

\item{min_per_subgroup}{minimum number of haplotypes per block per subgroup (default: 0)}

\item{subgroup_exception}{allow for not all subgroups to include a block}

\item{consider_all}{If TRUE always haplotypes which are not in the node to be in a generated block}

\item{save_allblock}{If TRUE keep all haplotypes with all windows according to a block (even under min_similarity)}

\item{block_extending}{If TRUE use the window-extending algorithm (step V)}

\item{max_extending_diff}{Maximum number of windows with different realisation in the block-extending-algorithm}

\item{extending_ratio}{Minimum Ratio between windows with one different realisation to multiple in block-extending-algorithm}

\item{min_majorblock_steps}{Number of steps till full filtering with min_majorblock is done}

\item{snp_extending}{If TRUE use the SNP-extending-algorithm (step V) (default: TRUE)}

\item{max_extending_diff_snp}{Maximum number of SNPs with variants in SNP-extending-algorithm (step V; default: 0)}

\item{extending_ratio_snp}{Minimum ratio of SNPs with only one allele to those with variants (default: Inf)#'}

\item{major_snp_calculation}{If TRUE calculate for major allele for each SNP in each block (default:TRUE)}

\item{off_node_addition}{If TRUE identify additional blocks in regions not covered by the window cluster (default: FALSE)}

\item{off_node_minimum_blocklength}{Minimum length of newly identified blocks (default: 10)}

\item{off_node_minimum_blocksize}{Minimum number of individuals in newly identified blocks (default: 5)}

\item{raster}{Raster-width in the identification step (default: 5; recommended to be lower than off_node_minimum_blocklength)}

\item{at_least_one}{If TRUE no allowed merging errors in windows of size 1}

\item{prefilter}{If TRUE filter the dataset before detecting for block-structure (default: FALSE)}

\item{maf}{Minimum minor allel frequency in prefilter (default: 0.05)}

\item{equal_remove}{If TRUE filter out SNPs in perfect correlation to the next SNP (default: FALSE)}

\item{big_output}{If TRUE return not only blocklist but also blockinfo, dataset ect.}

\item{blockinfo_mode}{Structure of the groups in step I (default: 0 - Common haplos as major variants, 1- minimum number of groups)}

\item{c_dhm_mode}{If TRUE use high speed calculation with C in block_merging (default: TRUE)}

\item{intersect_func}{Selection of intersect function used in the computation (default: 2 - efficient C, 1- efficient R, 0- base::intersect)}

\item{fast_compiler}{If TRUE use compiler-package to enableJIT(3) (default: TRUE)}

\item{max_groups}{Maximum number of groups per window (adaptive window-size, default: 0 - use fixed window size)}

\item{recoding}{If TRUE change allele coding (Major allele "A", Minor allele "C")}

\item{recoding_notneeded}{Set to TRUE if dataset only contains "A", "C" to save comp. time}

\item{consider_multi}{If TRUE use multi-level edges to identify blocks (default: FALSE)}

\item{multi_min}{minimum number of haplotypes per multi transition to use in consider_multi (default: 5)}

\item{blockinfo_mode_na}{Set TRUE for activation of NA-modelling}

\item{actual_snp_weight}{Set weight for difference between two alleles in a SNP (more than 1 possible base pair)}

\item{na_snp_weight}{Set weight for difference between NA and allele in a SNP (more than 1 possible base pair)}

\item{na_seq_weight}{Set weight for difference between NA and allele a loci with 1 possible base pair}

\item{weighting_length}{Weighting factor for length to determine major block (default: 1)}

\item{weighting_size}{Weighting factor for number of haplotypes in block to determine major block (default: 1)}

\item{recalculate_biggest}{Set to FALSE to only calculate the number of major positions for those blocks that could be removed in each iteration (number of major blocks is only increasing when removing other blocks)}

\item{target_coverage}{Target Coverage in the blocklist}

\item{max_iteration}{Maximum number of iterations to archive target coverage}

\item{min_step_size}{Minimum increase/decrase in min_block_size in target coverage fitting}

\item{target_stop}{Stop fitting target coverage fitting early}

\item{multi_window_mode}{Set TRUE so active the computation of multi window cluster with separate window_size/merging_error/min_share}

\item{adaptive_mode}{Set to TRUE to use the predefined adoptive mode ( window_size = c(5,10,20,50), target_coverage=0.9)}

\item{developer_mode}{Set to TRUE to not delete internal computational stuff in the blocklist}

\item{double_share}{Extended-block-identification share t}

\item{early_remove}{Remove Nodes with little number of haplotypes before SM, SG cycle (default: FALSE)}

\item{node_min_early}{Minimum number of haplotypes per node before SM, SG cycle (default: NULL)}

\item{min_reduction_cross}{Minimum number of nodes reduction to continue SM, SG cycle (default: -Inf, till fixation)}

\item{min_reduction_neglet}{Minimum number of nodes reduction to continue NN, SM, SG, SM cycle (default: -Inf, till fixation)}

\item{parallel_window}{Maximum number of markers to process jointly (default: Inf)}

\item{window_overlap}{Overlap between windows in parallel mode (default: 0)}

\item{window_cores}{Number of cores to use in parallel mode (default: 1)}
}
\description{
Function to generate haplotype blocks from haploid data
}
