% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_closeblock_merging.R
\name{block_closeblock_merging}
\alias{block_closeblock_merging}
\title{Merging neighbouring blocks}
\usage{
block_closeblock_merging(blocklist, blockinfo, indi, nwindow, max_diff_l,
  max_diff_i, intersect_func = intersect, dataset)
}
\arguments{
\item{blocklist}{block-dataset}

\item{blockinfo}{List with all relevant information to each window seperatly}

\item{indi}{number of haplotypes in the dataset}

\item{nwindow}{number of windows in the dataset}

\item{max_diff_l}{maximum number of windows with different haplotypes inbetween (default: 1)}

\item{max_diff_i}{maximum number of individuals in only one of the two blocks (default: 1)}

\item{intersect_func}{Used intersect-function (internally relevant for computation time)}

\item{dataset}{dataset which variant nr. for each window}
}
\value{
haplotype library
}
\description{
Function to merging neighbouring block with similar haplotypes/allels
}
