% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_dataset_construction.R
\name{block_dataset_construction}
\alias{block_dataset_construction}
\title{Construct blockinfo dataset}
\usage{
block_dataset_construction(
  blockinfo,
  blocklist = NULL,
  indi = NULL,
  nwindow = NULL
)
}
\arguments{
\item{blockinfo}{List with all relevant information to each window seperatly}

\item{blocklist}{blocklist created by blocklist_calculcation algorithm}

\item{indi}{number of haplotypes in the dataset}

\item{nwindow}{number of windows in the dataset}
}
\value{
Window-based dataset
}
\description{
Internal Function to generate a dataset containing which variant in present in with window
}
\examples{
data(ex_maze)
temp1 <- block_calculation(ex_maze, big_output=TRUE)
block_dataset <- block_dataset_construction(temp1[[4]][[1]], temp1[[1]])
}
