% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehh_block.R
\name{block_ehh}
\alias{block_ehh}
\title{Function to derive bEHH}
\usage{
block_ehh(blocklist = NULL, data = NULL, marker, plot = FALSE,
  position1 = NULL, standardization = 3, group = NULL,
  return_ehh = TRUE)
}
\arguments{
\item{blocklist}{blocklist}

\item{data}{haploid SNP-dataset}

\item{marker}{marker for which to calculate eHH/iHH}

\item{plot}{generate a bEHH-Plot (default: FALSE)}

\item{position1}{Position in bp (default: 1,2,3,...)}

\item{standardization}{Standardization by allele freq (0), allele freq of individuals in blocks (1), block (2), non (3)}

\item{group}{List containing all individuals assign in each group to derive different scores for subgroups}

\item{return_ehh}{Return single ehh values}
}
\description{
Function to generate haplotype blocks from haploid data
}
