% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ihh_block.R
\name{block_ihh}
\alias{block_ihh}
\title{Mainfunction to calculate haplotype blocks}
\usage{
block_ihh(
  blocklist = NULL,
  data = NULL,
  plot = FALSE,
  position1 = NULL,
  standardization = 3,
  group = NULL,
  verbose = TRUE
)
}
\arguments{
\item{blocklist}{blocklist}

\item{data}{haploid SNP-dataset}

\item{plot}{generate a IHH-Plot (default: FALSE)}

\item{position1}{Position in bp (default: 1,2,3,...)}

\item{standardization}{Standardization by allele freq (0), allele freq of individuals in blocks (1), block (2), non (3)}

\item{group}{List containing all individuals assign in each group to derive different scores for subgroups}

\item{verbose}{Set to FALSE to not display any prints}
}
\value{
Block-based IHH scores
}
\description{
Function to generate haplotype blocks from haploid data
}
\examples{
data(blocklist_ex_maze)
ihh_scores <- block_ihh(blocklist_ex_maze, plot=TRUE)
}
