% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_dataset_construction.R
\name{block_matrix_construction}
\alias{block_matrix_construction}
\title{Construct blocklist dataset}
\usage{
block_matrix_construction(blocklist, indi = NULL)
}
\arguments{
\item{blocklist}{blocklist created by blocklist_calculcation algorithm}

\item{indi}{number of haplotypes in the dataset}
}
\value{
Haplotype block-based dataset
}
\description{
Internal Function to generate a dataset containing which variant in present in with window
}
\examples{
data(blocklist_ex_maze)
block_dataset <- block_matrix_construction(blocklist_ex_maze)
}
