% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_windowdataset.R
\name{block_windowdataset}
\alias{block_windowdataset}
\title{Function to derive block-dataset (uniform-window)}
\usage{
block_windowdataset(blocklist = NULL, data = NULL,
  consider_nonblock = FALSE, return_dataset = FALSE,
  non_haploblocker = FALSE)
}
\arguments{
\item{blocklist}{blocklist}

\item{data}{haploid SNP-dataset}

\item{consider_nonblock}{Consider haplotypes that are in no block (default: FALSE)}

\item{return_dataset}{Return a per window dataset with more than 2 variants instead of binary coded one (default: FALSE)}

\item{non_haploblocker}{Set TRUE to consider all haplotypes and all variants in each window}
}
\description{
Function to compute a block-dataset with uniform start/end points
}
