% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockinfo_biggest.R
\name{blockinfo_biggest}
\alias{blockinfo_biggest}
\title{Compute biggest block for each region}
\usage{
blockinfo_biggest(blocklist, nwindow = NULL, indi = NULL,
  type = "window", min_majorblock = 5000, weighting_length = 1,
  weighting_size = 1, recalculate_biggest = TRUE, window_size)
}
\arguments{
\item{blocklist}{block-dataset}

\item{nwindow}{number of windows in the dataset}

\item{indi}{number of haplotypes in the dataset}

\item{type}{length measure (default: "window" , alt: "snp", "bp")}

\item{min_majorblock}{minimum of positions in the dataset a block is the biggest covering (default: 50)}

\item{weighting_length}{Weighting factor for length to determine major block (default: 1)}

\item{weighting_size}{Weighting factor for number of haplotypes in block to determine major block (default: 1)}

\item{recalculate_biggest}{Set to FALSE to only calculate the number of major positions for those blocks that could be removed in each iteration (number of major blocks is only increasing when removing other blocks)}

\item{window_size}{size of each window in the algorithm (default: 20)}
}
\description{
Internal function to count positions where each block is major block
}
