% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockinfo_calculation.R
\name{blockinfo_calculation}
\alias{blockinfo_calculation}
\title{Mainfunction to calculate haplotype blocks}
\usage{
blockinfo_calculation(
  dhm,
  window_sequence = NULL,
  window_size = 20,
  merging_error = 1,
  bp_map = NULL,
  window_anchor_gens = NULL,
  at_least_one = TRUE,
  max_groups = 0,
  blockinfo_mode = 0,
  c_dhm = NULL,
  verbose = TRUE
)
}
\arguments{
\item{dhm}{haploid SNP-dataset}

\item{window_sequence}{sequence of predefined windows (default: NULL ;per row: start$snp, end$snp, length, length - merging_error, start$bp, end$bp)}

\item{window_size}{size of each window in the algorithm (default: 20)}

\item{merging_error}{number of allowed errors per block (default: 1)}

\item{bp_map}{vector of positions for each SNP in bp (default: NULL - all 0)}

\item{window_anchor_gens}{matrix to constructed window_sequence base on start/end points in bp (e.g. gen regions, per row: start, end)}

\item{at_least_one}{If TRUE no allowed merging errors in windows of size 1}

\item{max_groups}{Maximum number of groups per window (adaptive window-size, default: 0 - use fixed window size)}

\item{blockinfo_mode}{Structure of the groups in step I (default: 0 - Common haplos as major variants, 1- minimum number of groups)}

\item{c_dhm}{Bit-wise coded SNP-dataset}

\item{verbose}{Set to FALSE to not display any prints}
}
\value{
[[1]] window-based haplotype information [[2]] window sequence
}
\description{
Function to generate haplotype blocks from haploid data
}
