% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockinfo_calculation_na.R
\name{blockinfo_calculation_na}
\alias{blockinfo_calculation_na}
\title{Mainfunction to calculate haplotype blocks}
\usage{
blockinfo_calculation_na(dhm, window_sequence = NULL, window_size = 32,
  merging_error = 3, bp_map = NULL, window_anchor_gens = NULL,
  at_least_one = TRUE, blockinfo_mode = 0, actual_snp_weight = 5,
  na_snp_weight = 2, na_seq_weight = 0, verbose = TRUE)
}
\arguments{
\item{dhm}{haploid SNP-dataset}

\item{window_sequence}{sequence of predefined windows (default: NULL ;per row: start$snp, end$snp, length, length - merging_error, start$bp, end$bp)}

\item{window_size}{size of each window in the algorithm (default: 20)}

\item{merging_error}{number of allowed errors per block (default: 1)}

\item{bp_map}{vector of positions for each SNP in bp (default: NULL - all 0)}

\item{window_anchor_gens}{matrix to constructed window_sequence base on start/end points in bp (e.g. gen regions, per row: start, gen)}

\item{at_least_one}{If TRUE no allowed merging errors in windows of size 1}

\item{blockinfo_mode}{Structure of the groups in step I (default: 0 - Common haplos as major variants, 1- minimum number of groups)}

\item{actual_snp_weight}{Set weight for difference between two alleles in a SNP (more than 1 possible base pair)}

\item{na_snp_weight}{Set weight for difference between NA and allele in a SNP (more than 1 possible base pair)}

\item{na_seq_weight}{Set weight for difference between NA and allele a loci with 1 possible base pair}

\item{verbose}{Set to FALSE to not display any prints}
}
\value{
[[1]] window-based haplotype information [[2]] window sequence
}
\description{
Function to generate haplotype blocks from haploid data
}
