% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockinfo_major.R
\name{blockinfo_major}
\alias{blockinfo_major}
\title{Old Major Version Identifier}
\usage{
blockinfo_major(blockinfo, dhm, window_size)
}
\arguments{
\item{blockinfo}{List with all relevant information to each window seperatly}

\item{dhm}{haploid SNP-dataset}

\item{window_size}{size of each window in the algorithm (default: 20)}
}
\value{
window-based haplotype information
}
\description{
Old internal function to identify the major version of each haplotype in each window
}
