% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockinfo_size.R
\name{blockinfo_max}
\alias{blockinfo_max}
\title{Determine the maximum number of the same haplotype in each window}
\usage{
blockinfo_max(blockinfo)
}
\arguments{
\item{blockinfo}{List with all relevant information to each window seperatly}
}
\value{
Maximum number of same variants in each window
}
\description{
Function to determine the number of the same haplotype in each window
}
