% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockinfo_size.R
\name{blockinfo_size0}
\alias{blockinfo_size0}
\title{Determine number of differnt haplotypes per window}
\usage{
blockinfo_size0(blockinfo)
}
\arguments{
\item{blockinfo}{List with all relevant information to each window seperatly}
}
\value{
Number of variants (before merging) in each window
}
\description{
Function to determine number of differnt haplotypes per window (before error reduction)
}
