% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocklist_comparison.R
\name{blocklist_comparison}
\alias{blocklist_comparison}
\title{Comparison of blocklists}
\usage{
blocklist_comparison(blocklist1, blocklist2, indi = NULL, type = "snp",
  compair_region = NULL, shift = 0, turn.around = FALSE,
  intersect_func = intersect)
}
\arguments{
\item{blocklist1}{blocklist created by blocklist_calculcation algorithm}

\item{blocklist2}{blocklist created by blocklist_calculcation algorithm}

\item{indi}{number of haplotypes in the dataset}

\item{type}{length measure (default: "window" , alt: "snp", "bp")}

\item{compair_region}{Region to compair (default: Every overlapping SNP)}

\item{shift}{switch the positions of the second blocklist (default: 0)}

\item{turn.around}{If TRUE turn the results of blocklist2 around}

\item{intersect_func}{Used intersect-function (internally relevant for computation time)}
}
\value{
similarity score/matrix between the two blocklist
}
\description{
Function compair different blocklists
}
\examples{
data(ex_maze)
data(blocklist_ex_maze)
blocklist2 <- block_calculation(ex_maze, min_majorblock = 1000)
temp1 <- blocklist_comparison(blocklist_ex_maze, blocklist2)
}
