% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex_maze.R
\name{blocklist_ex_maze}
\alias{blocklist_ex_maze}
\title{Blocklist of the ex_maze dataset}
\source{
HaploBlocker output
}
\usage{
blocklist_ex_maze
}
\description{
A dataset of 313 DH-lines of the first chromosome of maize
}
\author{
Torsten Pook \email{torsten.pook@uni-goettingen.de}
}
