% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocklist_reorder.R
\name{blocklist_reorder}
\alias{blocklist_reorder}
\title{Sort Blocklist}
\usage{
blocklist_reorder(blocklist, node_min = 5)
}
\arguments{
\item{blocklist}{block-dataset}

\item{node_min}{minimum number of haplotypes per block (default: 5)}
}
\value{
haplotype library
}
\description{
Function to sort the blocks based on starting-position
}
