% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocklist_size.R
\name{blocklist_size}
\alias{blocklist_size}
\title{Calculate size of each block}
\usage{
blocklist_size(blocklist, intersect_func = intersect, first_block = 1)
}
\arguments{
\item{blocklist}{block-dataset}

\item{intersect_func}{Used intersect-function (internally relevant for computation time)}

\item{first_block}{First block to consider in the computation (default: 1)}
}
\value{
Number of haplotypes in each block
}
\description{
Function to calculate the size of each block
}
\examples{
data(blocklist_ex_maze)
blocklist_size(blocklist_ex_maze)
}
