% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocklist_startend.R
\name{blocklist_startend}
\alias{blocklist_startend}
\title{Merge blocks}
\usage{
blocklist_startend(blocklist, type = "snp", first_block = 1, internal = FALSE)
}
\arguments{
\item{blocklist}{block-dataset}

\item{type}{length measure (default: "window" , alt: "snp", "bp")}

\item{first_block}{First block to consider in the computation (default: 1)}

\item{internal}{Set to TRUE to now provide col/rownames for start/end points}
}
\value{
start/end points of haplotype blocks
}
\description{
Function to merge/remove/modify blocks
}
\examples{
data(blocklist_ex_maze)
blocklist_startend(blocklist_ex_maze)
}
