% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocklist_subset.R
\name{blocklist_subset}
\alias{blocklist_subset}
\title{Merge blocks}
\usage{
blocklist_subset(blocklist, subset)
}
\arguments{
\item{blocklist}{block-dataset}

\item{subset}{Subset of haplotype to reduce the haplotype library to}
}
\value{
haplotype library
}
\description{
Function to merge/remove/modify blocks
}
