% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex_maze.R
\name{C_decodeSNPs}
\alias{C_decodeSNPs}
\title{Internal variable}
\source{
Internal variable
}
\usage{
C_decodeSNPs
}
\description{
Internal variable from C
}
\author{
Torsten Pook \email{torsten.pook@uni-goettingen.de}
}
