% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_new_transition.R
\name{calculate_new_transition}
\alias{calculate_new_transition}
\title{Simple Merge nodes}
\usage{
calculate_new_transition(data, nodes, nwindow, add_e = NULL,
  add_a = NULL, intersect_func = intersect)
}
\arguments{
\item{data}{node-dataset}

\item{nodes}{node for which to calculate transition probabilities}

\item{nwindow}{number of windows in the dataset}

\item{add_e}{additional possible previous nodes}

\item{add_a}{additional possible following nodes}

\item{intersect_func}{Used intersect-function (internally relevant for computation time)}
}
\value{
window cluster node with updated transition probabilities
}
\description{
Internal Function to calculate transition probabilities for a single node
}
