% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_transition.R
\name{calculate_transition}
\alias{calculate_transition}
\title{Calculate transition probabilities between nodes}
\usage{
calculate_transition(data1, intersect_func = intersect)
}
\arguments{
\item{data1}{node-dataset}

\item{intersect_func}{Used intersect-function (internally relevant for computation time)}
}
\value{
window cluster
}
\description{
Internal Function to calculate transition probabilities between nodes
}
