% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_test.R
\name{coverage_test}
\alias{coverage_test}
\title{Coverage Test}
\usage{
coverage_test(blocklist, indi = NULL, type = "snp", max = 1)
}
\arguments{
\item{blocklist}{block-dataset}

\item{indi}{number of haplotypes in the dataset}

\item{type}{length measure (default: "window" , alt: "snp", "bp")}

\item{max}{maximum number in each cell (overlapping blocks; default: 1)}
}
\value{
coverage matrix
}
\description{
Function to calculate the coverage of the blocks over the whole dataset
}
\examples{
data(blocklist_ex_maze)
t <- coverage_test(blocklist_ex_maze)
}
