% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossmerge.R
\name{crossmerge}
\alias{crossmerge}
\title{Cross Merge nodes}
\usage{
crossmerge(data, indi, nwindow, a = NULL, intersect_func = intersect)
}
\arguments{
\item{data}{node-dataset}

\item{indi}{number of haplotypes in the dataset}

\item{nwindow}{number of windows in the dataset}

\item{a}{Internal helper for a efficient abort criterion}

\item{intersect_func}{Used intersect-function (internally relevant for computation time)}
}
\description{
Internal Function perform a cross merge for node data
}
