% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{data_import}
\alias{data_import}
\title{Function to import genomic data from a vcf/ped file}
\usage{
data_import(dhm, inbred = FALSE, verbose = TRUE)
}
\arguments{
\item{dhm}{Path of the file to import}

\item{inbred}{Set to TRUE when working with inbred material and importing genomic data from file}
}
\value{
genomic dataset and basepair positions from the vcf/ped-file
}
\description{
Function to import genomic data from a vcf/ped file
}
