% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upper_estimator.R
\name{diploid_upper}
\alias{diploid_upper}
\title{Library presents diploid}
\usage{
diploid_upper(blocklist, s0m, min_similarity = 0.99)
}
\arguments{
\item{blocklist}{block-dataset}

\item{s0m}{dataset of diploid individuals to check}

\item{min_similarity}{minimum rate of the same SNPs to be added to the block (default: 0.99)}
}
\value{
Frequency estimated for each haplotype block
}
\description{
Calculate upper limited of present in a blocklist (ignore possible phasing errors)
}
