% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upper_estimator.R
\name{diploid_upper2}
\alias{diploid_upper2}
\title{Library presents diploid}
\usage{
diploid_upper2(blocklist, s0m, min_similarity = 0.99)
}
\arguments{
\item{blocklist}{block-dataset}

\item{s0m}{dataset of diploid individuals to check}

\item{min_similarity}{XXX}
}
\description{
Calculate upper limited of present in a blocklist (ignore possible phasing errors)
}
