% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_block.R
\name{extend_block}
\alias{extend_block}
\title{Extend-Block}
\usage{
extend_block(
  blocklist,
  indi,
  nwindow,
  max_extending_diff = 1,
  extending_ratio = 10,
  dataset,
  window_sequence_list
)
}
\arguments{
\item{blocklist}{blocklist created by blocklist_calculcation algorithm}

\item{indi}{number of haplotypes in the dataset}

\item{nwindow}{number of windows in the dataset}

\item{max_extending_diff}{Maximum number of windows with different realisation in the block-extending-algorithm}

\item{extending_ratio}{Minimum Ratio between windows with one different realisation to multiple in block-extending-algorithm}

\item{dataset}{dataset which variant nr. for each window}

\item{window_sequence_list}{sequence of predefined windows (default: NULL ;per row: start$snp, end$snp, length, length - merging_error, start$bp, end$bp)}
}
\value{
haplotype library
}
\description{
Function to add additional windows to each block
}
