% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_snp.R
\name{extend_snp}
\alias{extend_snp}
\title{Extend-SNP}
\usage{
extend_snp(blocklist, indi, nwindow, dhm, window_sequence_list, bp_map,
  max_extending_diff_snp = 0, extending_ratio_snp = Inf)
}
\arguments{
\item{blocklist}{blocklist created by blocklist_calculcation algorithm}

\item{indi}{number of haplotypes in the dataset}

\item{nwindow}{number of windows in the dataset}

\item{dhm}{haploid SNP-dataset}

\item{window_sequence_list}{sequence of predefined windows (default: NULL ;per row: start$snp, end$snp, length, length - merging_error, start$bp, end$bp)}

\item{bp_map}{vector of positions for each SNP in bp (default: NULL - all 0)}

\item{max_extending_diff_snp}{Maximum number of SNPs with variants in SNP-extending-algorithm (step V; default: 0)}

\item{extending_ratio_snp}{Minimum ratio of SNPs with only one allele to those with variants (default: Inf)}
}
\value{
haplotype library
}
\description{
Function to add single SNPs to each block
}
