% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/founder_detection.R
\name{founder_detection}
\alias{founder_detection}
\title{Detection of founder segments}
\usage{
founder_detection(
  dhm = NULL,
  dhm_founder = NULL,
  dhm_off = NULL,
  founding = NULL,
  plot_perline = FALSE,
  plot_overview = TRUE,
  lines_to_plot = 1:50,
  min_concordance = 0,
  multicolor = FALSE,
  err = 0.025,
  err2 = 0.1,
  rec = 1e-04,
  ped = NULL,
  weighting_length = 2,
  weighting_size = 1,
  target_coverage = NULL,
  min_majorblock = NULL,
  big_output = FALSE
)
}
\arguments{
\item{dhm}{haploid SNP-dataset (founders need to be in the first few columns)}

\item{dhm_founder}{haploid SNP-dataset of the offspring}

\item{founding}{Number of founding haplotypes in the population}

\item{plot_perline}{Generate an per line plot of estimated founders}

\item{plot_overview}{Generate a plot of the lines chosen in lines_to_plot}

\item{multicolor}{Set to TRUE to indicate regions with multiple potentialy founder haplotypes by use of multiple colors}

\item{err}{Hidden Markov parameter (chance of error within a haplotype block; default: 0.025}

\item{err2}{Hidden Markov parameter (chance of error outside of haplotype blocks; default: 0.1)}

\item{rec}{Hidden Markov parameter (chance of a recombination event between two markers; default: 0.001)}

\item{ped}{Pedigree - this will lead to lines only having founders from the lines given ancestors}

\item{weighting_length}{Weighting factor for length to determine major block (default: 1)}

\item{weighting_size}{Weighting factor for number of haplotypes in block to determine major block (default: 1)}

\item{target_coverage}{Target Coverage in the blocklist}

\item{min_majorblock}{minimum of cells in the dataset a block is the biggest covering (default: 5000)}

\item{big_output}{If TRUE return not only blocklist but also blockinfo, dataset ect.}

\item{min_condorance}{Minimum concordance of a segment to be assigned to a founder (default: 0; everything will be assigned)}
}
\value{
Estimated Founder for each cell in the haploid dataset (dhm)
}
\description{
Function to add off-variant-blocks
}
