% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_blocks.R
\name{identify_blocks}
\alias{identify_blocks}
\title{Identify possible blocks base on the node-dataset}
\usage{
identify_blocks(data, indi, nwindow, min_share = 0.95,
  consider_nodes = TRUE, consider_edge = TRUE, edge_min = 5,
  subgroups = NULL, min_per_subgroup = 0, intersect_func = intersect,
  identify_filter = FALSE, consider_multi = FALSE, multi_min = 5,
  subgroup_exception = 0, double_share = 1, node_min = 5)
}
\arguments{
\item{data}{node-dataset}

\item{indi}{number of haplotypes in the dataset}

\item{nwindow}{number of windows in the dataset}

\item{min_share}{minimum percentage of transition to the same block for extension (default: 0.95, step III)}

\item{consider_nodes}{Use nodes to identify blocks (default: TRUE)}

\item{consider_edge}{Use edges between nodes to identify blocks (default: TRUE)}

\item{edge_min}{minimum number of haplotypes per transition to use in consider_edge (default: 5)}

\item{subgroups}{possible subgroups to consider in the block identification process (default: NULL - list(1:indi))}

\item{min_per_subgroup}{minimum number of haplotypes per block per subgroup (default: 0)}

\item{intersect_func}{Used intersect-function (internally relevant for computation time)}

\item{identify_filter}{If TRUE filter out Starting blocks to avoid duplicated (default: FALSE)}

\item{consider_multi}{If TRUE use multi-level edges to identify blocks (default: FALSE)}

\item{multi_min}{minimum number of haplotypes per multi transition to use in consider_multi (default: 5)}

\item{subgroup_exception}{allow for not all subgroups to include a block}

\item{double_share}{Extended-block-identification share t}

\item{node_min}{minimum number of haplotypes per block (default: 5)}
}
\value{
haplotype library
}
\description{
Function to identify possible blocks base on the node-dataset
}
