% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ignore_small_nodes.R
\name{ignore_small_nodes}
\alias{ignore_small_nodes}
\title{Ignore small nodes}
\usage{
ignore_small_nodes(data, indi, nwindow, node_min = 5, gap = 10,
  intersect_func = intersect)
}
\arguments{
\item{data}{node-dataset}

\item{indi}{number of haplotypes in the dataset}

\item{nwindow}{number of windows in the dataset}

\item{node_min}{minimum number of haplotypes per block (default: 5)}

\item{gap}{remove haplotypes in nodes adjacent to nodes with less than minimim_blocksize haplotypes in it (default: 10 windows)}

\item{intersect_func}{Used intersect-function (internally relevant for computation time)}
}
\value{
window cluster
}
\description{
Internal Function remove windows which transition to less frequent haplotypes in adjecent windows
}
