% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indi_calc.R
\name{indi_calc}
\alias{indi_calc}
\title{Number of Indi}
\usage{
indi_calc(blocklist)
}
\arguments{
\item{blocklist}{block-dataset}
}
\value{
Number of individuals considered in the haplotype library
}
\description{
Function to determine number of individuals in a blocklist
}
\examples{
data(blocklist_ex_maze)
indi_calc(blocklist_ex_maze)
}
