% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llength.R
\name{llength}
\alias{llength}
\title{Help function for list length}
\usage{
llength(list)
}
\arguments{
\item{list}{List}
}
\value{
Length of each list element
}
\description{
Function to derive the length of lists
}
\examples{
data(blocklist_ex_maze)
llength(blocklist_ex_maze)
}
