% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/major_snp_calculation.R
\name{major_snp_calculation}
\alias{major_snp_calculation}
\title{Calculate Major SNPs}
\usage{
major_snp_calculation(blocklist, dhm, recoding = FALSE)
}
\arguments{
\item{blocklist}{blocklist created by blocklist_calculcation algorithm}

\item{dhm}{haploid SNP-dataset}

\item{recoding}{If TRUE change allele coding (Major allele "A", Minor allele "C")}
}
\value{
haplotype library with entered allele variants / frequencies
}
\description{
Function to calculate the common allele for each SNP of a block
}
