% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes_size.R
\name{nodes_size}
\alias{nodes_size}
\title{nodes size}
\usage{
nodes_size(data)
}
\arguments{
\item{data}{node-dataset}
}
\value{
size of each node in the window cluster
}
\description{
Function to calculate the number of haplotypes in each node
}
