% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_removal.R
\name{overlap_removal}
\alias{overlap_removal}
\title{Function to remove overlapping block segments}
\usage{
overlap_removal(blocklist = NULL, data = NULL, node_min = 5)
}
\arguments{
\item{blocklist}{blocklist}

\item{data}{window cluster (third output in block_calculation when big_output is set to TRUE)}

\item{node_min}{minimum number of haplotypes per block (default: 5)}
}
\value{
haplotype library with no overlapping blocks
}
\description{
Function to remove overlapping block segments
}
\examples{
data(ex_maze)
temp1 <- block_calculation(ex_maze, big_output=TRUE)
blocklist <- overlap_removal(temp1[[1]], temp1[[3]])
}
