% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_block.R
\name{plot_block}
\alias{plot_block}
\title{Plot Block_structure}
\usage{
plot_block(blocklist, type = "snp", orientation = "snp",
  include = TRUE, indi = NULL, min_to_plot = 5, intensity = 0.7,
  add_sort = TRUE, max_step = 500, snp_ori = NULL,
  export_order = FALSE, import_order = FALSE, xlim = NULL,
  n_colors = 20)
}
\arguments{
\item{blocklist}{block-dataset}

\item{type}{length measure (default: "window" , alt: "snp", "bp")}

\item{orientation}{Position on with to sort haplotypes for similarity (default: "snp", options: "front", "mid","back")}

\item{include}{Include haplotypes in no block in the orientation position in the graph}

\item{indi}{number of haplotypes in the dataset}

\item{min_to_plot}{Only include blocks with at least this many haplotypes in it (default:5)}

\item{intensity}{Color-intensity for the plotted blocks (default: 0.5 (options: 0-1))}

\item{add_sort}{If FALSE deactivate sorting haplotype for local similarity}

\item{max_step}{Maximum distance of blocks to consider in the orientation step (default: 500)}

\item{snp_ori}{Select SNP for the orientation algorithm if orientation=="snp"}

\item{export_order}{Export the chosen haplotype order after sorting}

\item{import_order}{Import a given haplotype order}

\item{xlim}{X-axis boundaries of positions to include in the plot}

\item{n_colors}{Number of different colors used for visualization (without package RColorBrewer limited to 8)}
}
\value{
Visualization of the haplotype library
}
\description{
Function to create a graphical representation of the blocklist
}
\examples{
data(blocklist_ex_maze)
plot_block(blocklist_ex_maze)
}
