% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renaming_combi.R
\name{renaming_combi}
\alias{renaming_combi}
\title{Renaming function}
\usage{
renaming_combi(data, nwindow)
}
\arguments{
\item{data}{node-dataset}

\item{nwindow}{number of windows in the dataset}
}
\value{
Updated window cluster
}
\description{
Internal Function to remove empty nodes and rename after merging
}
