% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_vcf.R
\name{write_vcf}
\alias{write_vcf}
\title{Function to remove overlapping block segments}
\usage{
write_vcf(
  blocklist = NULL,
  path = NULL,
  chromo = 1,
  type = "bp",
  hetero = FALSE
)
}
\arguments{
\item{blocklist}{blocklist}

\item{path}{path to write the vcf-file to}

\item{chromo}{Chromosome at hand (default: 1)}

\item{type}{Position in the vcf file (when available base-bair, alt: "snp", "window")}

\item{hetero}{Set to TRUE when working on diploid species. Two chromosome sets are assumed to be next to each other in dhm}
}
\value{
VCF-file for the blocklist
}
\description{
Function to remove overlapping block segments
}
\examples{
data(ex_maze)
blocklist <- block_calculation(ex_maze, bp=1:9999)
write_vcf(blocklist, path=tempdir())
}
