% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upper_estimator.R
\name{haploid_upper}
\alias{haploid_upper}
\title{Library presents haploid}
\usage{
haploid_upper(blocklist, s0m, min_similarity = 0.99)
}
\arguments{
\item{blocklist}{block-dataset}

\item{s0m}{dataset of diploid individuals to check}

\item{min_similarity}{XXX}
}
\description{
Calculate upper limited of present in a blocklist (not-ignore possible phasing errors)
}
