% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_removal.R
\name{overlap_removal}
\alias{overlap_removal}
\title{Function to remove overlapping block segments}
\usage{
overlap_removal(blocklist = NULL, data = NULL)
}
\arguments{
\item{blocklist}{blocklist}

\item{data}{window cluster (third output in block_calculation when big_output is set to TRUE)}
}
\description{
Function to remove overlapping block segments
}
